create folders components and pages for movie, refactory movie-routing
