+++ libs/movie/src/index.ts
@@ -3,9 +3,9 @@ export * from './lib/movie/+state';
 export * from './lib/movie/guards/movie-active.guard';
 export * from './lib/movie/guards/movie-list.guard';
 export * from './lib/movie/movie.module';
-export { ViewComponent } from './lib/movie/view/view.component';
+export { MovieViewComponent } from './lib/movie/pages/movie-view/movie-view.component';
 export { ContainerComponent } from './lib/movie/form/container.component';
-export { ListComponent } from './lib/movie/list/list.component';
+export { MovieListComponent } from './lib/movie/pages/movie-list/movie-list.component';
 export { HomeComponent } from './lib/movie/home/home.component';
 export { default as staticModels } from './lib/movie/staticModels';
