+++ libs/movie/src/lib/movie/+state/movie.model.ts
@@ -1,6 +1,6 @@
 import { Organization } from "@blockframes/organization";
 import { Material } from "@blockframes/material";
-import { Stakeholder } from "@blockframes/movie/stakeholder/+state";
+import { Stakeholder } from "../../stakeholder/+state";
 
 export interface Movie {
   id: string,
