+++ libs/movie/src/lib/movie/components/home-empty/home-empty.component.ts
@@ -7,6 +7,4 @@ import { Component, ChangeDetectionStrategy } from '@angular/core';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class HomeEmptyComponent {
-
-  constructor() { }
 }
