+++ libs/movie/src/lib/movie/form/container.component.html
@@ -8,7 +8,7 @@
         <mat-icon>arrow_left</mat-icon>
       </ng-template>
     </button>
-    <movie-financing-view [form$]="form$" mode="preview"></movie-financing-view>
+    <movie-view [form$]="form$" mode="preview"></movie-view>
   </mat-drawer-content>
   <mat-drawer opened #drawer mode="side" position="end" [ngClass]="{ 'mat-drawer-fullscreen' : fullScreen }">
     <button mat-icon-button class="toggle-fullscreen" (click)="toggleFullScreen()">
