+++ libs/movie/src/lib/movie/pages/movie-list/movie-list.component.html
