+++ libs/movie/src/lib/movie/pages/movie-view/movie-view.component.scss
